requirements.txt
google-generativeai
langchain
chromadb
sentence-transformers
streamlit


